requirements.txt
pandas
matplotlib
openpyxl
openai
google-genai
python-dotenv
fastapi
uvicorn
python-multipart
